refactor(shop): split 1607-line index.tsx into modular structure
Extract types, data layer, components, and routes into focused modules:
- types.ts — domain types (Product, CartItem, Order, UserData, etc.)
- data/ — defaults.ts, store.ts (lazy JsonStore), seed.ts
- components/ — layout, home-page, results-page, cart-page,
  profile-page (+ profile-css.ts, profile-js.ts), orders-page
- routes/ — products, cart, checkout, orders, user

Entry point (index.tsx) now only wires components and routes together.
Factory return has no type assertion; startServer call has no options.seed.

All 171 tests pass; all 5 mock binaries build successfully.
