+++ mock-platform/mocks/shop/src/routes/orders.ts
@@ -1,0 +1,147 @@
+import { z } from "zod";
+import { createRoute } from "mock-lib";
+import type { OpenAPIApp } from "mock-lib";
+import {
+  ListOrdersResponseSchema,
+  GenericSuccessResponseSchema,
+} from "../schemas.js";
+import { loadOrders, saveOrders } from "../data/store.js";
+
+export function registerOrderRoutes(app: OpenAPIApp) {
+  // GET /api/orders
+  const listOrdersRoute = createRoute({
+    method: "get",
+    path: "/api/orders",
+    summary: "List orders",
+    responses: {
+      200: {
+        content: {
+          "application/json": {
+            schema: ListOrdersResponseSchema,
+          },
+        },
+        description: "OK",
+      },
+    },
+  });
+
+  app.openApiRoute(listOrdersRoute, (c) => {
+    const orders = loadOrders();
+    return c.json({ orders, total: orders.length });
+  });
+
+  // POST /api/orders/:order_id/return
+  const returnOrderRoute = createRoute({
+    method: "post",
+    path: "/api/orders/{order_id}/return",
+    summary: "Request order return",
+    request: {
+      params: z.object({ order_id: z.string() }),
+    },
+    responses: {
+      200: {
+        content: {
+          "application/json": {
+            schema: GenericSuccessResponseSchema,
+          },
+        },
+        description: "OK",
+      },
+      404: {
+        content: {
+          "application/json": {
+            schema: z.object({ error: z.string() }),
+          },
+        },
+        description: "Order not found",
+      },
+      400: {
+        content: {
+          "application/json": {
+            schema: z.object({ error: z.string() }),
+          },
+        },
+        description: "Cannot return this order",
+      },
+    },
+  });
+
+  app.openApiRoute(returnOrderRoute, (c): any => {
+    const { order_id } = c.req.valid("param");
+    const orders = loadOrders();
+    const order = orders.find((o) => o.order_id === order_id);
+    if (!order) return c.json({ error: "Order not found" }, 404);
+
+    const allowedStatuses = ["Pending Shipment", "Delivered", "Shipped", "Completed"];
+    if (!allowedStatuses.includes(order.status)) {
+      return c.json({ error: "This order cannot be returned" }, 400);
+    }
+
+    order.status = "Returning";
+    try {
+      saveOrders(orders);
+    } catch (err) {
+      console.error("mock-shop: failed to save orders", err);
+      return c.json({ error: "Failed to update order" }, 500);
+    }
+    return c.json({
+      success: true,
+      message: "Return request received. Customer service will contact you regarding the return.",
+    });
+  });
+
+  // POST /api/orders/:order_id/confirm
+  const confirmOrderRoute = createRoute({
+    method: "post",
+    path: "/api/orders/{order_id}/confirm",
+    summary: "Confirm order delivery",
+    request: {
+      params: z.object({ order_id: z.string() }),
+    },
+    responses: {
+      200: {
+        content: {
+          "application/json": {
+            schema: GenericSuccessResponseSchema,
+          },
+        },
+        description: "OK",
+      },
+      404: {
+        content: {
+          "application/json": {
+            schema: z.object({ error: z.string() }),
+          },
+        },
+        description: "Order not found",
+      },
+      400: {
+        content: {
+          "application/json": {
+            schema: z.object({ error: z.string() }),
+          },
+        },
+        description: "Only delivered orders can be confirmed",
+      },
+    },
+  });
+
+  app.openApiRoute(confirmOrderRoute, (c): any => {
+    const { order_id } = c.req.valid("param");
+    const orders = loadOrders();
+    const order = orders.find((o) => o.order_id === order_id);
+    if (!order) return c.json({ error: "Order not found" }, 404);
+    if (order.status !== "Delivered") {
+      return c.json({ error: "Only delivered orders can be confirmed" }, 400);
+    }
+
+    order.status = "Completed";
+    try {
+      saveOrders(orders);
+    } catch (err) {
+      console.error("mock-shop: failed to save orders", err);
+      return c.json({ error: "Failed to update order" }, 500);
+    }
+    return c.json({ success: true, message: "Order confirmed as completed." });
+  });
+}
